comments and restructuring
