+++ ebu_tt_live/scripts/ebu_simple_producer.py
@@ -1,15 +1,15 @@
 
 from itertools import cycle
-from nltk.tokenize import PunktSentenceTokenizer, BlanklineTokenizer, WhitespaceTokenizer
 from datetime import timedelta
 from twisted.internet import task, reactor, interfaces
 from zope.interface import implementer
 from twisted.python import log
 import logging
 from argparse import ArgumentParser
+from .common import create_loggers, tokenize_english_document
 
 from ebu_tt_live.clocks.local import LocalMachineClock
-from ebu_tt_live.streaming import BroadcastServerFactory, StreamingServerProtocol
+from ebu_tt_live.streaming import BroadcastServerFactory as wsFactory, StreamingServerProtocol
 from ebu_tt_live.documents import EBUTT3Document, TimeBase
 from ebu_tt_live.bindings import div_type, p_type, br_type
 
@@ -85,66 +85,11 @@ class SimplePullDocumentProducer(object):
         pass
 
 
-def tokenize_english_document(input_text):
-    """
-    This is a crude tokenizer for input conversations in English.
-    :param input_text:
-    :return:
-    """
-    end_list = []
-    block_tokenizer = BlanklineTokenizer()
-    sentence_tokenizer = PunktSentenceTokenizer()
-    word_tokenizer = WhitespaceTokenizer()
-    # using the 38 characters in one line rule from ITV subtitle guidelines
-    characters_per_line = 38
-    lines_per_subtitle = 2
 
-    blocks = block_tokenizer.tokenize(input_text)
-    for block in blocks:
-        # We have one speaker
-        sentences = sentence_tokenizer.tokenize(block)
-        # We have the sentences
-        for sentence in sentences:
-            words = word_tokenizer.tokenize(sentence)
-            reverse_words = words[::-1]
-
-            lines = []
-            current_line = ''
-            line_full = False
-            while reverse_words:
-                word = reverse_words.pop()
-                longer_line = ' '.join([current_line, word]).strip()
-                if len(longer_line) > characters_per_line and len(current_line):
-                    # The longer line is overreaching boundaries
-                    reverse_words.append(word)
-                    line_full = True
-                elif len(word) >= characters_per_line:
-                    # Very long words
-                    current_line = longer_line
-                    line_full = True
-                else:
-                    current_line = longer_line
-
-                if line_full:
-                    lines.append(current_line)
-                    current_line = ''
-                    line_full = False
-
-                if len(lines) >= lines_per_subtitle:
-                    end_list.append(lines)
-                    lines = []
-            if current_line:
-                lines.append(current_line)
-            if lines:
-                end_list.append(lines)
-
-    return end_list
 
 
 def main():
-    log_observer = log.PythonLoggingObserver(loggerName='twisted')
-    log_observer.start()
-    logging.basicConfig(level=logging.INFO)
+    create_loggers()
 
     parsed_args = parser.parse_args()
 
@@ -156,14 +101,12 @@ def main():
             while True:
                 yield [reference_clock.get_full_clock_time()]
 
-        subtitle_tokens = gen_time()
+        subtitle_tokens = gen_time()  # Instead of text we provide the availability time as content.
     else:
         with open('blargh.txt', 'r') as infile:
             full_text = infile.read()
 
-        subtitle_tokens = cycle(tokenize_english_document(full_text))
-
-    wsFactory = BroadcastServerFactory
+        subtitle_tokens = cycle(tokenize_english_document(full_text))  # This makes the source cycle infinitely.
 
     factory = wsFactory(u"ws://127.0.0.1:9000")
